+++ src/py42/sdk/queries/__init__.py
@@ -1,3 +1,4 @@
+from py42 import settings
 from py42.sdk.queries.query_filter import FilterGroup
 
 
@@ -5,17 +6,17 @@ class BaseQuery(object):
     def __init__(self, *args, **kwargs):
         self._filter_group_list = list(args)
         self._group_clause = kwargs.get(u"group_clause", u"AND")
-        self.page_size = 10000
+        self.page_number = kwargs.get(u"page_number") or 1
+        self.page_size = kwargs.get(u"page_size") or settings.security_events_per_page
         self.sort_direction = u"asc"
 
         # Override
-        self.page_number = None
         self.sort_key = None
 
     @classmethod
-    def from_dict(cls, _dict, group_clause=u"AND"):
+    def from_dict(cls, _dict, group_clause=u"AND", **kwargs):
         filter_groups = [FilterGroup.from_dict(item) for item in _dict[u"groups"]]
-        return cls(*filter_groups, group_clause=group_clause)
+        return cls(*filter_groups, group_clause=group_clause, **kwargs)
 
     @classmethod
     def any(cls, *args):