+++ src/py42/services/savedsearch.py
@@ -32,28 +32,32 @@ class SavedSearchService(BaseService):
         uri = u"{}/{}".format(self._resource, search_id)
         return self._connection.get(uri)
 
-    def get_query(self, search_id):
+    def get_query(self, search_id, page_number=None, page_size=None):
         """Get the saved search in form of a query(`py42.sdk.queries.fileevents.file_event_query`).
 
         Args:
             search_id (str): Unique search Id of the saved search.
+            page_number (int, optional): The consecutive group of results of size page_size in the result set to return. Defaults to None.
+            page_size (int, optional): The maximum number of results to be returned. Defaults to None.
         Returns:
             :class:`py42.sdk.queries.fileevents.file_event_query.FileEventQuery`
         """
         response = self.get_by_id(search_id)
         search = response[u"searches"][0]
-        return FileEventQuery.from_dict(search)
+        return FileEventQuery.from_dict(
+            search, page_number=page_number, page_size=page_size,
+        )
 
-    def execute(self, search_id, pg_num=1, pg_size=10000):
+    def execute(self, search_id, page_number=None, page_size=None):
         """
         Execute a saved search for given search Id and return its results.
 
         Args:
             search_id (str): Unique search Id of the saved search.
-            pg_num (int, optional): The consecutive group of results of size pg_size in the result set to return. Defaults to 1.
-            pg_size (int, optional): The maximum number of results to be returned. Defaults to 10,000.
+            page_number (int, optional): The consecutive group of results of size page_size in the result set to return. Defaults to None.
+            page_size (int, optional): The maximum number of results to be returned. Defaults to None.
         Returns:
             :class:`py42.response.Py42Response`
         """
-        query = self.get_query(search_id)
+        query = self.get_query(search_id, page_number=page_number, page_size=page_size)
         return self._file_event_client.search(query)