+++ CHANGELOG.md
@@ -35,6 +35,13 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 
 ### Added
 
+- Added additional user-adjustable setting for security events page size:
+    - `py42.settings.security_events_per_page`
+
+- Page page_number and page_size parameters for saved search queries:
+    - `py42.securitydata.savedsearches.get_query()`
+    - `py42.securitydata.savedsearches.execute()`
+
 - Methods for obtaining archive information:
     - `sdk.archive.get_by_archive_guid`
     - `sdk.archive.get_all_by_device_guid`